% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.location}
\alias{as.location}
\alias{as.location.file}
\alias{as.location.url}
\title{Convert a path or URL to a location object.}
\usage{
as.location(x, ...)

\method{as.location}{file}(x, ...)

\method{as.location}{url}(x, ...)
}
\arguments{
\item{x}{Input.}

\item{...}{Ignored.}
}
\description{
Convert a path or URL to a location object.
}

