% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conversions.R
\name{as.SpatialPolygonsDataFrame}
\alias{as.SpatialPolygonsDataFrame}
\title{Convert to various sp classes from geojson files}
\usage{
as.SpatialPolygonsDataFrame(x, ...)
}
\arguments{
\item{x}{Input.}

\item{...}{Further args passed on to \code{\link[rgdal]{readOGR}}}
}
\description{
Convert to various sp classes from geojson files
}

