% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{capturedWriteOGR}
\alias{capturedWriteOGR}
\title{An \code{writeOGR} shim to capture stdout}
\usage{
capturedWriteOGR(obj, obj_size, layer, writeOGR)
}
\arguments{
\item{obj}{the \code{Spatial} object to use}

\item{obj_size}{pls pass in \code{object.size(obj)}}

\item{layer}{spatial layer to use}

\item{writeOGR}{pls pass in \code{writeOGR} (no quotes)}
}
\value{
character vector of GeoJSON if all goes well
}
\description{
An \code{writeOGR} shim to capture stdout
}
\examples{
\dontrun{
capturedWriteOGR(cities[1:10,],
                 object.size(cities[1:10,]),
                 "cities",
                 writeOGR))
}
}

