% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwgr.R
\name{cwgr}
\alias{cwgr}
\title{capture writeOGR}
\usage{
cwgr(obj, layer = "", layer_options = "")
}
\arguments{
\item{obj}{Object, e.g, a data.frame}

\item{layer}{layer name, default: \code{""}}

\item{layer_options}{layer options, as a character vector}
}
\description{
capture writeOGR
}
\examples{
\dontrun{
library("rgdal")
cities <- readOGR(system.file("vectors", package = "rgdal")[1], "cities")
cwgr(cities[1:10,], "cities")

coordinates(states) <- ~long+lat
cwgr(states[1:100,], "states")
}
}

