% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_topo.R
\name{geo2topo}
\alias{geo2topo}
\alias{topo2geo}
\alias{topo2geo.character}
\alias{topo2geo.default}
\title{GeoJSON to TopoJSON and back}
\usage{
geo2topo(x, ...)

topo2geo(x, ...)

\method{topo2geo}{default}(x, ...)

\method{topo2geo}{character}(x, ...)
}
\arguments{
\item{x}{geojson or topojson as a character string, or form a file, or url}

\item{...}{ignored}
}
\value{
A character string of either GeoJSON or TopoJSON
}
\description{
TopoJSON to GeoJSON not quite working yet.
}
\examples{
# geojson to topojson
x <- '{"type": "LineString", "coordinates": [ [100.0, 0.0], [101.0, 1.0] ]}'
cat(geo2topo(x))
}

