% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geojson_read.R
\name{geojson_read}
\alias{geojson_read}
\title{Read geojson from a local file or a URL}
\usage{
geojson_read(x, query, method = "web", parse = FALSE, what = "list", ...)
}
\arguments{
\item{x}{Path to a local file or a URL.}

\item{query}{A SQL query.}

\item{method}{One of web or local. Matches on partial strings.}

\item{parse}{(logical) To parse geojson to data.frame like structures if possible.
Default: \code{FALSE}}

\item{what}{What to return. One of list or sp (for Spatial class). Default: list.
If sp chosen, method forced to be "local".}

\item{...}{Ignored}
}
\description{
Read geojson from a local file or a URL
}
\details{
Uses \code{\link{file_to_geojson}} internally.

This function supports various geospatial file formats from a URL, as well as local
kml, shp, and geojson file formats.
}
\examples{
\dontrun{
# From a file
file <- system.file("examples", "california.geojson", package = "geojsonio")
out <- geojson_read(file)

# From a URL
url <- "https://raw.githubusercontent.com/glynnbird/usstatesgeojson/master/california.geojson"
geojson_read(url, method = "local")

# Use as.location first if you want
geojson_read(as.location(file))

# use jsonlite to parse to data.frame structures where possible
geojson_read(url, method = "local", parse = TRUE)

# output a SpatialClass object
## read kml
file <- system.file("examples", "norway_maple.kml", package = "geojsonio")
geojson_read(as.location(file), what = "sp")
## read geojson
file <- system.file("examples", "california.geojson", package = "geojsonio")
geojson_read(as.location(file), what = "sp")
## read geojson from a url
url <- "https://raw.githubusercontent.com/glynnbird/usstatesgeojson/master/california.geojson"
geojson_read(url, what = "sp")

# Query PostGIS/Postgres
library("DBI")
library("RPostgres")
conn <- dbConnect(RPostgres::Postgres(), dbname = 'postgistest')
state <- "SELECT row_to_json(fc)
 FROM (SELECT 'FeatureCollection' As type, array_to_json(array_agg(f)) As features
 FROM (SELECT 'Feature' As type
    , ST_AsGeoJSON(lg.geog)::json As geometry
    , row_to_json((SELECT l FROM (SELECT loc_id, loc_name) As l
      )) As properties
   FROM locations As lg   ) As f )  As fc;"
json <- geojson_read(conn, state)[[1]]
map_leaf(json)

# doesn't work right now
## file <- system.file("examples", "feature_collection.geojson", package = "geojsonio")
## geojson_read(file, what = "sp")
}
}

