% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{geojsonio-defunct}
\alias{geojsonio-defunct}
\title{Defunct functions in geojsonio}
\description{
\itemize{
\item \code{\link[=lint]{lint()}}: See \code{geojsonlint::geojson_hint}
\item \code{\link[=validate]{validate()}}: See \code{geojsonlint::geojson_lint}
}
}
