% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojsonio-package.r
\name{geojsonio-deprecated}
\alias{geojsonio-deprecated}
\title{Deprecated functions in geojsonio}
\description{
\itemize{
 \item \code{\link{lint}}: In the next version this function will be removed, defunct.
 See \code{geojsonlint::geojson_hint}
 \item \code{\link{validate}}: In the next version this function will be removed, defunct.
 See \code{geojsonlint::geojson_lint}
}
}
