% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geojsonio-package.R
\docType{package}
\name{geojsonio}
\alias{geojsonio}
\alias{geojsonio-package}
\title{Convert various data formats to/from GeoJSON or TopoJSON}
\description{
This package focuses mostly on converting lists, data.frame's, numeric, SpatialPolygons,
SpatialPolygonsDataFrame, and more to GeoJSON with the help of \code{rgdal} and
friends. You can currently read TopoJSON - writing TopoJSON will come in a future
version of this package.
}
\author{
Scott Chamberlain

Andy Teucher
}

