% Generated by roxygen2 (4.0.2): do not edit by hand
\name{map_gist}
\alias{map_gist}
\title{Publish an interactive map as a GitHub gist}
\usage{
map_gist(geojsonfile, description = "", public = TRUE, browse = TRUE, ...)
}
\arguments{
\item{geojsonfile}{Path to a geojson file}

\item{description}{Description for the Github gist, or leave to default (=no description)}

\item{public}{(logical) Want gist to be public or not? Default: TRUE}

\item{browse}{If TRUE (default) the map opens in your default browser.}

\item{...}{Further arguments passed on to \code{\link[httr]{POST}}}
}
\description{
There are two ways to authorize to work with your GitHub account:

\itemize{
 \item PAT - Generate a personal access token (PAT) at
 \url{https://help.github.com/articles/creating-an-access-token-for-command-line-use} and
 record it in the GITHUB_PAT envar in your \code{.Renviron} file.
 \item Interactive - Interactively login into your GitHub account and authorise with OAuth.
}

Using the PAT method is recommended.

Using the gist_auth() function you can authenticate seperately first, or if you're not
authenticated, this function will run internally with each functionn call. If you have a
PAT, that will be used, if not, OAuth will be used.
}
\examples{
\donttest{
url <- 'http://www.zillow.com/static/shp/ZillowNeighborhoods-MT.zip'
file <- file_to_geojson(input=url, method='web', outfilename='zillow_mt')
map_gist(file)
}
}

