% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgis.R
\name{postgis}
\alias{postgis}
\title{PostGIS setup}
\description{
\code{\link{geojson_read}} allows you to get data out of a PostgreSQL 
database set up with PostGIS. Below are steps for setting up data
that we can at the end query with \code{\link{geojson_read}}.
}
\details{
If you don't already have PostgreSQL or PostGIS:
\itemize{
 \item PostgreSQL installation: http://www.postgresql.org/download/
 \item PostGIS installation: http://postgis.net/install
}

Once you have both of those installed, you can proceed below.
}
