% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pretty.R
\name{pretty}
\alias{pretty}
\title{Convert json input to pretty printed output}
\usage{
pretty(x, indent = 4)
}
\arguments{
\item{x}{Input, character string}

\item{indent}{(integer) Number of spaces to indent}
}
\description{
Convert json input to pretty printed output
}
\details{
Only works with json class input. This is a simple wrapper around
\code{\link[jsonlite]{prettify}}, so you can easily use that yourself.
}

