% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geojsonio-package.r
\docType{data}
\name{states}
\alias{states}
\title{This is the same data set from the ggplot2 library}
\description{
This is a data.frame with "long", "lat", "group", "order", "region", and "subregion"
columns specifying polygons for each US state.
}
\keyword{data}

