% Generated by roxygen2 (4.0.1): do not edit by hand
\name{style_geojson}
\alias{style_geojson}
\alias{style_geojson.data.frame}
\alias{style_geojson.list}
\title{Style a data.frame prior to converting to geojson.}
\usage{
style_geojson(input, var = NULL, var_col = NULL, var_sym = NULL,
  var_size = NULL, color = NULL, symbol = NULL, size = NULL)

\method{style_geojson}{data.frame}(input, var = NULL, var_col = NULL,
  var_sym = NULL, var_size = NULL, color = NULL, symbol = NULL,
  size = NULL)

\method{style_geojson}{list}(input, var = NULL, var_col = NULL,
  var_sym = NULL, var_size = NULL, color = NULL, symbol = NULL,
  size = NULL)
}
\arguments{
\item{input}{A data.frame or a list}

\item{var}{A single variable to map colors, symbols, and/or sizes to.}

\item{var_col}{The variable to map colors to.}

\item{var_sym}{The variable to map symbols to.}

\item{var_size}{The variable to map size to.}

\item{color}{Valid RGB hex color}

\item{symbol}{An icon ID from the Maki project \url{http://www.mapbox.com/maki/} or
a single alphanumeric character (a-z or 0-9).}

\item{size}{One of 'small', 'medium', or 'large'}
}
\description{
Style a data.frame prior to converting to geojson.
}
\examples{
\dontrun{
## Style geojson - from data.frames
library("RColorBrewer")
library("maps")
data(us.cities)
smalluscities <-
   subset(us.cities, country.etc == 'OR' | country.etc == 'NY' | country.etc == 'CA')

### Just color
style_geojson(smalluscities, var = 'country.etc',
   color=brewer.pal(length(unique(smalluscities$country.etc)), "Blues"))
### Just size
style_geojson(smalluscities, var = 'country.etc', size=c('small','medium','large'))
### Color and size
style_geojson(smalluscities, var = 'country.etc',
   color=brewer.pal(length(unique(smalluscities$country.etc)), "Blues"),
   size=c('small','medium','large'))

## From lists
mylist <- list(list(latitude=30, longitude=120, state="US"),
               list(latitude=32, longitude=130, state="OR"),
               list(latitude=38, longitude=125, state="NY"),
               list(latitude=40, longitude=128, state="VT"))
style_geojson(mylist, var = 'state',
   color=brewer.pal(length(unique(sapply(mylist, '[[', 'state'))), "Blues"))
}
}

