% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to_geo_json}
\alias{as.json}
\alias{as.json.geo_list}
\alias{as.json.geojson}
\alias{to_geo_json}
\alias{to_geo_json.SpatialPointsDataFrame}
\alias{to_geo_json.SpatialPolygons}
\alias{to_geo_json.SpatialPolygonsDataFrame}
\alias{to_geo_json.data.frame}
\alias{to_geo_json.list}
\alias{to_geo_json.numeric}
\title{Convert many input types with spatial data to geojson specified as a json string}
\usage{
to_geo_json(...)

\method{to_geo_json}{SpatialPolygons}(input, ...)

\method{to_geo_json}{SpatialPolygonsDataFrame}(input, ...)

\method{to_geo_json}{SpatialPointsDataFrame}(input, ...)

\method{to_geo_json}{numeric}(input, polygon = NULL)

\method{to_geo_json}{data.frame}(input, lat = "latitude", lon = "longitude",
  polygon = NULL, ...)

\method{to_geo_json}{list}(input, lat = "latitude", lon = "longitude",
  polygon = NULL, ...)

as.json(x, ...)

\method{as.json}{geo_list}(x, ...)

\method{as.json}{geojson}(x, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[jsonlite]{toJSON}}}

\item{input}{Input list, data.frame, or spatial class. Inputs can also be dplyr \code{tbl_df}
class since it inherits from \code{data.frame}.}

\item{polygon}{If a polygon is defined in a data.frame, this is the column that defines the
grouping of the polygons in the \code{data.frame}}

\item{lat}{Latitude name. Default: latitude}

\item{lon}{Longitude name. Default: longitude}
}
\description{
Convert many input types with spatial data to geojson specified as a json string
}
\details{
This function creates a geojson structure as a json character string; it does not
write a file using \code{rgdal} - see \code{\link{to_geojson}} for that.
}
\examples{
\dontrun{
# From a numeric vector of length 2
to_geo_json(c(32.45,-99.74))

# From a data.frame
library('maps')
data(us.cities)
to_geo_json(us.cities[1:2,], lat='lat', lon='long')

# From SpatialPolygons class
library('sp')
poly1 <- Polygons(list(Polygon(cbind(c(-100,-90,-85,-100),
   c(40,50,45,40)))), "1")
poly2 <- Polygons(list(Polygon(cbind(c(-90,-80,-75,-90),
   c(30,40,35,30)))), "2")
sp_poly <- SpatialPolygons(list(poly1, poly2), 1:2)
to_geo_json(sp_poly)
to_geo_json(sp_poly, pretty=TRUE)
to_geo_json(sp_poly, pretty=TRUE, auto_unbox=TRUE)

# data.frame to SpatialPolygonsDataFrame
library('maps')
data(us.cities)
to_geojson(us.cities[1:2,], lat='lat', lon='long') \%>\% as.SpatialPolygonsDataFrame

# data.frame to json (via SpatialPolygonsDataFrame)
to_geojson(us.cities[1:2,], lat='lat', lon='long') \%>\% as.json
}
}

