% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to_geojson}
\alias{as.geoson}
\alias{to_geojson}
\alias{to_geojson.SpatialPointsDataFrame}
\alias{to_geojson.SpatialPolygons}
\alias{to_geojson.SpatialPolygonsDataFrame}
\alias{to_geojson.data.frame}
\alias{to_geojson.list}
\alias{to_geojson.numeric}
\title{Convert many input types with spatial data to a geojson file}
\usage{
to_geojson(...)

\method{to_geojson}{SpatialPolygons}(input, file = "myfile.geojson", ...)

\method{to_geojson}{SpatialPolygonsDataFrame}(input, file = "myfile.geojson",
  ...)

\method{to_geojson}{SpatialPointsDataFrame}(input, file = "myfile.geojson",
  ...)

\method{to_geojson}{numeric}(input, lat = "latitude", lon = "longitude",
  polygon = NULL, file = "myfile.geojson", ...)

\method{to_geojson}{data.frame}(input, lat = "latitude", lon = "longitude",
  polygon = NULL, file = "myfile.geojson", ...)

\method{to_geojson}{list}(input, lat = "latitude", lon = "longitude",
  polygon = NULL, file = "myfile.geojson", ...)

as.geoson(path, type)
}
\arguments{
\item{...}{Further args passed on to \code{\link[rgdal]{writeOGR}}}

\item{input}{Input list, data.frame, or spatial class. Inputs can also be dplyr \code{tbl_df}
class since it inherits from \code{data.frame}.}

\item{file}{A path and file name (e.g., myfile), with the .geojson on the end.}

\item{lat}{Latitude name. Default: latitude}

\item{lon}{Longitude name. Default: longitude}

\item{polygon}{If a polygon is defined in a data.frame, this is the column that defines the
grouping of the polygons in the \code{data.frame}}
}
\description{
Convert many input types with spatial data to a geojson file
}
\examples{
\dontrun{
# From a data.frame
library('maps')
data(us.cities)
to_geojson(us.cities[1:2,], lat='lat', lon='long')

# From polygons in R
library('ggplot2')
states <- map_data("state")
head(states)
to_geojson(input=states, lat='lat', lon='long', polygon='group')

## From a list
mylist <- list(list(latitude=30, longitude=120, marker="red"),
               list(latitude=30, longitude=130, marker="blue"))
to_geojson(mylist)

# From a numeric vector of length 2
vec <- c(32.45,-99.74)
to_geojson(vec)

# From SpatialPolygons class
library('sp')
poly1 <- Polygons(list(Polygon(cbind(c(-100,-90,-85,-100),
   c(40,50,45,40)))), "1")
poly2 <- Polygons(list(Polygon(cbind(c(-90,-80,-75,-90),
   c(30,40,35,30)))), "2")
sp_poly <- SpatialPolygons(list(poly1, poly2), 1:2)
to_geojson(sp_poly)

# From SpatialPolygonsDataFrame class
sp_polydf <- as(sp_poly, "SpatialPolygonsDataFrame")
to_geojson(input = sp_polydf, output = "file")
to_geojson(input = sp_polydf, output = "file", file = "~/stuff")
}
}
\seealso{
\code{\link{to_geo_list}}, \code{\link{to_geo_json}}
}

