% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{togeojson}
\alias{togeojson}
\alias{togeojson-package}
\title{Convert various data formats to geoJSON or topoJSON}
\description{
This package converts lists, data.frame's, numeric, SpatialPolygons, and
SpatialPolygonsDataFrame to geoJSON with the help of \code{rgdal} and friends, or topoJSON
with the help of the npm node package \code{topojson}
}
\author{
Scott Chamberlain
}

