% Generated by roxygen2 (4.0.2): do not edit by hand
\name{togeojson_create_gist}
\alias{togeojson_create_gist}
\title{Function that takes a list of files and creates payload for API}
\usage{
togeojson_create_gist(filenames, description = "", public = TRUE)
}
\arguments{
\item{filenames}{names of files to post}

\item{description}{brief description of gist (optional)}

\item{public}{whether gist is public (defaults to TRUE)}
}
\description{
Function that takes a list of files and creates payload for API
}
\keyword{internal}

