% Generated by roxygen2 (4.0.2): do not edit by hand
\name{topojson_read}
\alias{topojson_read}
\alias{topojson_read.character}
\alias{topojson_read.location}
\title{Read topojson from a local file or a URL}
\usage{
topojson_read(...)

\method{topojson_read}{character}(x, ...)

\method{topojson_read}{location}(x, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[rgdal]{readOGR}}}

\item{x}{Path to a local file or a URL.}
}
\description{
Read topojson from a local file or a URL
}
\examples{
\donttest{
# From a file
topojson_read("~/zillow_or.topojson")

# From a URL
url <- "https://raw.githubusercontent.com/shawnbot/d3-cartogram/master/data/us-states.topojson"
topojson_read(url)

# Use as.location first if you want
topojson_read(as.location("~/zillow_or.geojson"))
}
}

