% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geojsonio-package.r
\docType{data}
\name{us_cities}
\alias{us_cities}
\title{This is the same data set from the maps library, named differently}
\description{
This is the same data set from the maps library, named differently
}
\keyword{data}

