% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate-defunct}
\alias{validate-defunct}
\alias{validate}
\title{Validate a geoJSON file, json object, list, or Spatial class.}
\usage{
validate(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Validate a geoJSON file, json object, list, or Spatial class.
}
\keyword{internal}
