% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geojsonio-package.r
\docType{data}
\name{world_cities}
\alias{world_cities}
\title{This is the same data set from the maps library, named differently}
\format{A list with 6 components, namely "name", "country.etc", "pop", "lat", "long", and
"capital", containing the city name, the country name, approximate population
(as at January 2006), latitude, longitude and capital status indication (0 for non-capital,
1 for capital, 2 for China Municipalities, and 3 for China Provincial capitals)}
\description{
This database is primarily of world cities of population greater than about 40,000. Also
included are capital cities of any population size, and many smaller towns.
}
\keyword{data}

