% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.json.R
\name{as.json}
\alias{as.json}
\title{Convert inputs to JSON}
\usage{
as.json(x, ...)
}
\arguments{
\item{x}{Input}

\item{...}{Further args passed on to \code{\link[jsonlite]{toJSON}}}
}
\description{
Convert inputs to JSON
}

