% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.location}
\alias{as.location}
\alias{as.location.character}
\alias{as.location.location}
\alias{print.location}
\title{Convert a path or URL to a location object.}
\usage{
as.location(x, ...)

\method{as.location}{character}(x, ...)

\method{as.location}{location}(x, ...)

\method{print}{location}(x, ...)
}
\arguments{
\item{x}{Input.}

\item{...}{Ignored.}
}
\description{
Convert a path or URL to a location object.
}
\examples{
as.location("~/zillow_or.geojson")
url <- "https://gist.githubusercontent.com/sckott/5353e9522a4866729e63/raw/820939552c9dd7cfb7a4dab806e37ee973771533/pleiades245e5de1a252.geojson"
as.location(url)
}

