% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.location}
\alias{as.location}
\alias{as.location.character}
\alias{as.location.location}
\alias{print.location}
\title{Convert a path or URL to a location object.}
\usage{
as.location(x, ...)

\method{as.location}{character}(x, ...)

\method{as.location}{location}(x, ...)

\method{print}{location}(x, ...)
}
\arguments{
\item{x}{Input.}

\item{...}{Ignored.}
}
\description{
Convert a path or URL to a location object.
}
\examples{
# A file
file <- system.file("examples", "zillow_or.geojson", package = "togeojson")
as.location(file)

# A URL
url <- "https://raw.githubusercontent.com/glynnbird/usstatesgeojson/master/california.geojson"
as.location(url)
}

