% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conversions}
\alias{as.SpatialPolygonsDataFrame}
\alias{as.SpatialPolygonsDataFrame.geojson}
\alias{conversions}
\title{Convert to various sp classes from geojson files}
\usage{
as.SpatialPolygonsDataFrame(x, ...)

\method{as.SpatialPolygonsDataFrame}{geojson}(x, ...)
}
\description{
Bla bla
}

