% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conversions.R
\name{conversions}
\alias{as.SpatialPolygonsDataFrame}
\alias{as.SpatialPolygonsDataFrame.character}
\alias{as.SpatialPolygonsDataFrame.geojson}
\alias{conversions}
\title{Convert to various sp classes from geojson files}
\usage{
as.SpatialPolygonsDataFrame(x, ...)

\method{as.SpatialPolygonsDataFrame}{geojson}(x, ...)

\method{as.SpatialPolygonsDataFrame}{character}(x, ...)
}
\arguments{
\item{x}{Input.}

\item{...}{Further args passed on to \code{\link[rgdal]{readOGR}}}
}
\description{
Bla bla
}

