% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwgr.R
\name{cwgr}
\alias{cwgr}
\title{capture writeOGR}
\usage{
cwgr(obj, layer)
}
\description{
capture writeOGR
}
\examples{
\dontrun{
library("rgdal")
cities <- readOGR(system.file("vectors", package = "rgdal")[1], "cities")
cwgr(cities[1:10,], "cities")
}
}

