% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geojson_read}
\alias{geojson_read}
\alias{geojson_read.character}
\alias{geojson_read.location}
\title{Read geojson from a local file or a URL}
\usage{
geojson_read(...)

\method{geojson_read}{character}(x, ...)

\method{geojson_read}{location}(x, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[rgdal]{readOGR}}}

\item{x}{Path to a local file or a URL.}
}
\description{
Read geojson from a local file or a URL
}
\examples{
\donttest{
# From a file
file <- system.file("examples", "california.geojson", package = "togeojson")
out <- geojson_read(file)
plot(out)

# From a URL
url <- "https://raw.githubusercontent.com/glynnbird/usstatesgeojson/master/california.geojson"
geojson_read(url)
geojson_read(url, stringsAsFactors=FALSE)

# Use as.location first if you want
geojson_read(as.location(file))
}
}

