% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geojson_read}
\alias{geojson_read}
\alias{geojson_read.file}
\alias{geojson_read.url}
\title{Read a geojson file}
\usage{
geojson_read(...)

\method{geojson_read}{file}(input, ...)

\method{geojson_read}{url}(input, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[jsonlite]{toJSON}}}

\item{file}{Path to a local file or a URL.}
}
\description{
Read a geojson file
}
\examples{
\dontrun{
# From a file
geojson_read("~/asdfafaf")

# From a URL
geojson_read('<URL>')
}
}

