% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geojson_read}
\alias{geojson_read}
\alias{geojson_read.character}
\alias{geojson_read.location}
\title{Read geojson from a local file or a URL}
\usage{
geojson_read(...)

\method{geojson_read}{character}(x, ...)

\method{geojson_read}{location}(x, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[rgdal]{readOGR}}}

\item{x}{Path to a local file or a URL.}
}
\description{
Read geojson from a local file or a URL
}
\examples{
\donttest{
# From a file
geojson_read("~/zillow_or.geojson")

# From a URL
url <- "https://gist.githubusercontent.com/sckott/5353e9522a4866729e63/raw/820939552c9dd7cfb7a4dab806e37ee973771533/pleiades245e5de1a252.geojson"
geojson_read(url)
geojson_read(url, stringsAsFactors=FALSE)

# Use as.location first if you want
geojson_read(as.location("~/zillow_or.geojson"))
}
}

