% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geojson_read.R
\name{geojson_read}
\alias{geojson_read}
\title{Read geojson from a local file or a URL}
\usage{
geojson_read(x, ...)
}
\arguments{
\item{x}{Path to a local file or a URL.}

\item{...}{Further args passed on to \code{\link[rgdal]{readOGR}}}
}
\description{
Read geojson from a local file or a URL
}
\examples{
\donttest{
# From a file
file <- system.file("examples", "california.geojson", package = "togeojson")
out <- geojson_read(file)
plot(out)

# From a URL
url <- "https://raw.githubusercontent.com/glynnbird/usstatesgeojson/master/california.geojson"
geojson_read(url)
geojson_read(url, stringsAsFactors=FALSE)

# Use as.location first if you want
geojson_read(as.location(file))
}
}

