% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geojson_write}
\alias{as.geojson}
\alias{geojson_write}
\alias{geojson_write.SpatialPointsDataFrame}
\alias{geojson_write.SpatialPolygons}
\alias{geojson_write.SpatialPolygonsDataFrame}
\alias{geojson_write.data.frame}
\alias{geojson_write.list}
\alias{geojson_write.numeric}
\title{Convert many input types with spatial data to a geojson file}
\usage{
geojson_write(...)

\method{geojson_write}{SpatialPolygons}(input, file = "myfile.geojson", ...)

\method{geojson_write}{SpatialPolygonsDataFrame}(input,
  file = "myfile.geojson", ...)

\method{geojson_write}{SpatialPointsDataFrame}(input, file = "myfile.geojson",
  ...)

\method{geojson_write}{numeric}(input, lat = "latitude", lon = "longitude",
  polygon = NULL, file = "myfile.geojson", ...)

\method{geojson_write}{data.frame}(input, lat = "latitude",
  lon = "longitude", polygon = NULL, file = "myfile.geojson", ...)

\method{geojson_write}{list}(input, lat = "latitude", lon = "longitude",
  polygon = NULL, file = "myfile.geojson", ...)

as.geojson(path, type)
}
\arguments{
\item{...}{Further args passed on to \code{\link[rgdal]{writeOGR}}}

\item{input}{Input list, data.frame, or spatial class. Inputs can also be dplyr \code{tbl_df}
class since it inherits from \code{data.frame}.}

\item{file}{A path and file name (e.g., myfile), with the .geojson on the end.}

\item{lat}{Latitude name. Default: latitude}

\item{lon}{Longitude name. Default: longitude}

\item{polygon}{If a polygon is defined in a data.frame, this is the column that defines the
grouping of the polygons in the \code{data.frame}}

\item{path}{Path to file}

\item{type}{Type of file}
}
\description{
Convert many input types with spatial data to a geojson file
}
\examples{
\dontrun{
# From a data.frame
library('maps')
data(us.cities)
geojson_write(us.cities[1:2,], lat='lat', lon='long')

# From polygons in R
library('ggplot2')
states <- map_data("state")
head(states)
geojson_write(input=states, lat='lat', lon='long', polygon='group')

## From a list
mylist <- list(list(latitude=30, longitude=120, marker="red"),
               list(latitude=30, longitude=130, marker="blue"))
geojson_write(mylist)

# From a numeric vector of length 2
vec <- c(32.45,-99.74)
geojson_write(vec)

# From SpatialPolygons class
library('sp')
poly1 <- Polygons(list(Polygon(cbind(c(-100,-90,-85,-100),
   c(40,50,45,40)))), "1")
poly2 <- Polygons(list(Polygon(cbind(c(-90,-80,-75,-90),
   c(30,40,35,30)))), "2")
sp_poly <- SpatialPolygons(list(poly1, poly2), 1:2)
geojson_write(sp_poly)

# From SpatialPolygonsDataFrame class
sp_polydf <- as(sp_poly, "SpatialPolygonsDataFrame")
geojson_write(input = sp_polydf, output = "file")
geojson_write(input = sp_polydf, output = "file", file = "~/stuff")
}
}
\seealso{
\code{\link{geojson_list}}, \code{\link{geojson_json}}
}

