% Generated by roxygen2 (4.0.0): do not edit by hand
\name{map_gist}
\alias{map_gist}
\title{Publish an interactive map to view in the browser as a Github gist}
\usage{
map_gist(geojsonfile, description = "", file = "gistmap", browse = TRUE,
  ...)
}
\arguments{
\item{geojsonfile}{Path to a geojson file}

\item{description}{Description for the Github gist, or leave to default (=no description)}

\item{browse}{If TRUE (default) the map opens in your default browser.}

\item{...}{Further arguments passed on to \code{togeojson_gist}}
}
\value{
Creates a gist on your Github account, and prints out where the geojson file was
written on your machinee, the url for the gist, and an embed script in the console.
}
\description{
You will be asked ot enter you Github credentials (username, password) during
each session, but only once for each session. Alternatively, you could enter
your credentials into your .Rprofile file with the entries

\itemize{
 \item options(github.username = 'your_github_username')
 \item options(github.password = 'your_github_password')
}

then \code{map_gist} will simply read those options.

\code{map_gist} has modified code from the rCharts package by Ramnath Vaidyanathan
}
\examples{
\dontrun{
map_gist(geojsonfile='~/zillow_mt.geojson')
}
}

