% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pretty.R
\name{pretty}
\alias{pretty}
\title{Convert inputs to JSON}
\usage{
pretty(x, indent = 4)
}
\arguments{
\item{x}{Input, character string}

\item{indent}{(integer) Number of spaces to indent}
}
\description{
Convert inputs to JSON
}

