% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to_geojson}
\alias{to_geojson}
\alias{to_geojson.SpatialPolygons}
\alias{to_geojson.data.frame}
\alias{to_geojson.list}
\alias{to_geojson.numeric}
\title{Convert many input types with spatial data to geojson.}
\usage{
to_geojson(...)

\method{to_geojson}{SpatialPolygons}(input, lat = "latitude",
  lon = "longitude", polygon = NULL, output = "list", ...)

\method{to_geojson}{numeric}(input, lat = "latitude", lon = "longitude",
  polygon = NULL, output = "list", ...)

\method{to_geojson}{data.frame}(input, lat = "latitude", lon = "longitude",
  polygon = NULL, output = "list", ...)

\method{to_geojson}{list}(input, lat = "latitude", lon = "longitude",
  polygon = NULL, output = "list", ...)
}
\arguments{
\item{...}{Further args, not used.}

\item{input}{Input list, data.frame, or spatial class. Inputs can also be dplyr \code{tbl_df}
class since it inherits from \code{data.frame}.}

\item{lat}{Latitude name. Default: latitude}

\item{lon}{Longitude name. Default: longitude}

\item{polygon}{If a polygon is defined in a data.frame, this is the column that defines the
grouping of the polygons in the \code{data.frame}}

\item{output}{One of 'list' or 'geojson'. The output from a call to \code{to_geojson()}}
}
\description{
Includes support for lists, data.frame's, etc.
}
\details{
description...
}
\examples{
\dontrun{
library(maps)
data(us.cities)
head(us.cities)
to_geojson(input=us.cities, lat='lat', lon='long')

# polygons
library('ggplot2')
states <- map_data("state")
head(states)
## make list for input to e.g., rMaps
res <- to_geojson(input=states, lat='lat', lon='long', group='group')

## From a list
mylist <- list(list(latitude=30, longitude=120, marker="red"),
               list(latitude=30, longitude=130, marker="blue"))
to_geojson(mylist)

# From a numeric vector of length 2
vec <- c(32.45,-99.74)
to_geojson(vec)

# From SpatialPolygons class
poly1 <- Polygons(list(Polygon(cbind(c(32.45,-99.74,49.45,32.45),
   c(32.45,-99.74,32.45,32.45)))), "1")
poly2 <- Polygons(list(Polygon(cbind(c(39.45,-108.74,49.45,39.45),
   c(39.45,-99.74,32.45,39.45)))), "2")
sp_poly <- SpatialPolygons(list(poly1, poly2), 1:2)
to_geojson(sp_poly)
}
}

