\name{to_geojson}
\alias{to_geojson}
\alias{to_geojson.data.frame}
\alias{to_geojson.default}
\alias{to_geojson.list}
\alias{to_geojson.sp}
\title{Get palette actual name from longer names}
\usage{
to_geojson(...)

\method{to_geojson}{default}(input, lat = "latitude", lon = "longitude",
  ...)

\method{to_geojson}{data.frame}(input, lat = "latitude", lon = "longitude",
  ...)

\method{to_geojson}{list}(input, lat = "latitude", lon = "longitude", ...)

\method{to_geojson}{sp}(input, lat = "latitude", lon = "longitude", ...)
}
\arguments{
  \item{input}{}

  \item{lat}{Latitude name}

  \item{lon}{Longitude name}
}
\description{
Get palette actual name from longer names
}

