% Generated by roxygen2 (4.0.2): do not edit by hand
\name{topojson_read}
\alias{topojson_read}
\alias{topojson_read.file}
\alias{topojson_read.url}
\title{Read a topojson file}
\usage{
topojson_read(...)

\method{topojson_read}{file}(input, ...)

\method{topojson_read}{url}(input, ...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[jsonlite]{toJSON}}}

\item{file}{Path to a local file or a URL.}
}
\description{
Read a topojson file
}
\examples{
\dontrun{
# From a file
topojson_read("~/asdfafaf")

# From a URL
topojson_read('<URL>')
}
}

